% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aigitcraft-package.R
\name{AiGitcraft}
\alias{AiGitcraft}
\title{AI Gitcraft}
\description{
AiGitCraft is an R package designed to enhance Git operations by
leveraging Large Language Models (LLMs) to automatically generate meaningful
commit messages, pull request descriptions, and other related documentation.
}
