% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLM_operations.R
\name{describe_uncommitted_changes}
\alias{describe_uncommitted_changes}
\title{Describe the changes from the last committed state in a git repository.}
\usage{
describe_uncommitted_changes(
  repo_path = getOption("aigitcraft_repo", getwd()),
  screened_folders = NULL,
  staged = FALSE,
  use_description = TRUE,
  use_readme = TRUE,
  cite_changes = TRUE,
  suggest_commits = TRUE,
  ...
)
}
\arguments{
\item{repo_path}{The path to the repository.}

\item{screened_folders}{The path to the folders to be screened for changes.
If NULL (the default), all the folders in the repository are screened.}

\item{staged}{Logical indicating whether to include only staged changes.}

\item{use_description}{Logical indicating whether to include the content of
the DESCRIPTION file in the repository.}

\item{use_readme}{Logical indicating whether to include the content of the
README.md file in the repository.}

\item{cite_changes}{Logical indicating whether to include a prompt to cite
the relevant files and lines for the changes.}

\item{suggest_commits}{Logical indicating whether to include a prompt to
suggest a commit message and the involved files and lines for each logical
group of changes.}

\item{...}{Additional arguments forwarded to \code{\link[ellmer:chat-any]{ellmer::chat()}} to configure
the provider or model.}
}
\value{
A character string with the description of the changes.
}
\description{
This function uses LLM to describe the changes from the last committed state
in a git repository.
}
