% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLM_operations.R
\name{generate_twitter_thread}
\alias{generate_twitter_thread}
\title{Generate a Twitter thread}
\usage{
generate_twitter_thread(repo_path = getOption("aigitcraft_repo", getwd()))
}
\arguments{
\item{repo_path}{The path to the repository.}
}
\value{
A character string with the content of the Twitter thread.
}
\description{
This function uses LLM to generate a Twitter thread based on the repo README.md to present the repo to the community.
}
