% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_operations.R
\name{get_branch_differences}
\alias{get_branch_differences}
\title{Get the differences between two branches}
\usage{
get_branch_differences(
  repo_path = getOption("aigitcraft_repo", getwd()),
  target_branch = NULL,
  source_branch = "main",
  screened_folders = NULL
)
}
\arguments{
\item{repo_path}{The path to the repository.}

\item{target_branch}{The target branch (i.e. the branch you want to merge
from). Default is the current branch.}

\item{source_branch}{The source branch (i.e. the branch you want to merge
into). Default is 'main'.}

\item{screened_folders}{The path to the folders to be considered for changes.
If NULL (the default), all the folders in the repository are screened.}
}
\value{
A character string with the differences between the two branches,
displayed as a series of commits and their associated differences.
}
\description{
This function takes a repository path, a source branch, and a target branch
and returns the differences between the two branches.
}
