% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_operations.R
\name{get_uncommitted_changes}
\alias{get_uncommitted_changes}
\title{Get the uncommitted changes in the repository}
\usage{
get_uncommitted_changes(
  repo_path = getOption("aigitcraft_repo", getwd()),
  screened_folders = NULL,
  staged = FALSE
)
}
\arguments{
\item{repo_path}{The path to the repository.}

\item{screened_folders}{The path to the folders to be considered for changes.
If NULL (the default), all the folders in the repository are screened.}

\item{staged}{Whether to return the staged changes only. Default is FALSE.}
}
\value{
A character string with the uncommitted changes in the repository.
}
\description{
This function returns the uncommitted changes in the repository, staged or
unstaged.
}
