% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_repo_path}
\alias{validate_repo_path}
\title{Validate the repository path}
\usage{
validate_repo_path(repo_path)
}
\arguments{
\item{repo_path}{The path to the repository.}
}
\value{
Nothing.
}
\description{
This function validates the repository path provided by the user.
}
