% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLM_operations.R
\name{write_commit_message}
\alias{write_commit_message}
\title{Write a commit message for the staged changes}
\usage{
write_commit_message(
  repo_path = getOption("aigitcraft_repo", getwd()),
  use_conventional_commit = TRUE,
  use_description = TRUE,
  use_readme = TRUE,
  use_files = NULL,
  ...
)
}
\arguments{
\item{repo_path}{The path to the repository.}

\item{use_conventional_commit}{Logical indicating whether to style commit
messages according to the Conventional Commits specification
(https://www.conventionalcommits.org/en/v1.0.0/).}

\item{use_description}{Logical indicating whether to include the content of
the DESCRIPTION file to help understand the changes.}

\item{use_readme}{Logical indicating whether to include the content of the
README.md file to help understand the changes.}

\item{use_files}{A vector of file paths to analyze to help understand the
changes.}

\item{...}{Additional arguments forwarded to \code{\link[ellmer:chat-any]{ellmer::chat()}} to configure
the provider or model.}
}
\value{
A character string with the commit message.
}
\description{
This function uses LLM to write a commit message for the staged changes in a
git repository.
}
