% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_operations.R
\name{get_commit_differences}
\alias{get_commit_differences}
\title{Get the differences between two commits}
\usage{
get_commit_differences(
  target_commit = git2r::last_commit(getOption("aigitcraft_repo", getwd())),
  source_commit = NULL,
  screened_folders = NULL
)
}
\arguments{
\item{target_commit}{The target commit.}

\item{source_commit}{The source commit. If NULL, the parent commit of the
target commit is used.}

\item{screened_folders}{The path to the folders to be considered for changes.
If NULL (the default), all the folders in the repository are screened.}
}
\value{
A character string with the differences between the two commits. If
only the target commit is provided, then the output will be the details of
the target commit.
}
\description{
This function returns the differences between two commits.
}
