% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLM_operations.R
\name{write_pull_request_description}
\alias{write_pull_request_description}
\title{Create a pull request description automatically}
\usage{
write_pull_request_description(
  repo_path = getOption("aigitcraft_repo", getwd()),
  target_branch = git2r::repository_head()$name,
  source_branch = "main",
  screened_folders = NULL,
  use_description = TRUE,
  use_readme = TRUE,
  ...
)
}
\arguments{
\item{repo_path}{The path to the repository.}

\item{target_branch}{The target branch (i.e. the branch you want to merge
from). Default is the current branch.}

\item{source_branch}{The source branch (i.e. the branch you want to merge
into). Default is 'main'.}

\item{screened_folders}{The path to the folders to be considered for changes.
If NULL (the default), all the folders in the repository are screened.}

\item{use_description}{Logical indicating whether to include the content of
the DESCRIPTION file in the repository to help understand the changes.}

\item{use_readme}{Logical indicating whether to include the content of the
README.md file in the repository to help understand the changes.}

\item{...}{Additional arguments to be passed to \code{prompt_llm}.}
}
\value{
A character string with the pull request description.
}
\description{
This function uses LLM to create a pull request description automatically
using the differences between two branches.
}
